requirements.txt
itchat-uos
